// src/app/Interfaces/empresa.ts
export interface Empresa {
    id: number;
    anio: string;
    nombre: string;
    descripcion: string;
    rfc: string;
    pais: string;
    estado: string;
    cp: string;
    colonia: string;
    municipio: string;
    localidad: string;
    direccion: string;
    noExt: string;
    noInt: string;
    referencia: string;
    telefono: string;
    telefonoVenta: string;
    email: string;
    estatus: string;
    idCorporativo: string;
    corporativoDescripcion: string;
  }
  